+++ Geometry3D/render/renderer.py
@@ -1,15 +1,5 @@
-"""Abstract Renderer Module"""
-from .renderer_matplotlib import MatplotlibRenderer
-
-def Renderer(backend='matplotlib'):
-    """
-    **Input:**
-    
-    - backend: the backend of the renderer
-    
-    Only matplotlib is supported till now
-    """
-    if backend == 'matplotlib':
-        return MatplotlibRenderer()
-    else:
-        raise ValueError('Unknown backend %s' % (backend,))
+from ..visualization.visualizer import Visualizer
+from ..utils.logger import get_main_logger
+logger = get_main_logger()
+logger.warning("'Renderer' is deprecated, using 'Visualizer' instead.")
+Renderer = Visualizer