+++ Geometry3D/visualization/__init__.py
@@ -1,0 +1,4 @@
+from .visualizer import Visualizer
+
+
+__all__ = ('Visualizer',)