+++ changelog.md
@@ -10,3 +10,14 @@
 # 0.2.2
 - Major fix on spelling error on "polygon" which was "polygen".
 - Modify structure for files.
+
+# 0.2.3
+- Fix floating point precision problem.
+
+# 0.2.4
+- Fix floating point precision problem.
+
+# 0.3.0
+- Change the name of 'renderer' to 'visualizer'.
+- Add open3d backend for visualization. 
+- Fix known bugs.